refactor: ヘルスチェック関連のコードの位置を変更
